device/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
bidict==0.22.1
    # via python-socketio
click==8.1.6
    # via flask
flask==2.2.5
    # via
    #   -r requirements.in
    #   flask-socketio
flask-socketio==5.3.4
    # via -r requirements.in
importlib-metadata==6.7.0
    # via
    #   -r requirements.in
    #   flask
inky==1.5.0
    # via -r requirements.in
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
numpy==1.21.6
    # via
    #   -r requirements.in
    #   inky
pillow==9.5.0
    # via -r requirements.in
python-engineio==4.5.1
    # via python-socketio
python-socketio==5.8.0
    # via flask-socketio
rpi-gpio==0.7.1
    # via -r requirements.in
smbus2==0.4.2
    # via inky
spidev==3.6
    # via inky
werkzeug==2.2.3
    # via
    #   -r requirements.in
    #   flask
zipp==3.16.2
    # via importlib-metadata
